backend/requirements.txt
Django
django-nose==1.4.6
djangorestframework
nose==1.3.7
